+++ candidat-service/src/main/java/com/gogo/candidat_service/service/ExperienceServiceImpl.java
@@ -1,6 +1,8 @@
 package com.gogo.candidat_service.service;
 
 import com.gogo.candidat_service.dto.ExperienceDTO;
+import com.gogo.candidat_service.exception.CandidatNotFoundException;
+import com.gogo.candidat_service.exception.ExperienceNotFoundException;
 import com.gogo.candidat_service.mapper.ExperienceMapper;
 import com.gogo.candidat_service.model.Candidat;
 import com.gogo.candidat_service.model.Experience;
@@ -22,27 +24,38 @@ public class ExperienceServiceImpl implements ExperienceService {
     }
 
     @Override
-    public ExperienceDTO addExperience(Long candidatId, ExperienceDTO dto) {
+    public List<ExperienceDTO> addExperiences(Long candidatId, List<ExperienceDTO> dtos) throws CandidatNotFoundException {
         Candidat candidat = candidatRepository.findById(candidatId)
-                .orElseThrow(() -> new RuntimeException("Candidat non trouvé avec id " + candidatId));
+                .orElseThrow(() -> new CandidatNotFoundException("Candidat non trouvé avec id " + candidatId));
 
-        Experience experience = ExperienceMapper.fromDTO(dto, candidat);
+        List<Experience> experiences = dtos.stream()
+                .map(dto -> ExperienceMapper.fromDTO(dto, candidat))
+                .toList();
 
-        Experience saved = experienceRepository.save(experience);
-        return ExperienceMapper.toDTO(saved);
-    }
+        List<Experience> savedExperiences = experienceRepository.saveAll(experiences);
 
-    @Override
-    public List<ExperienceDTO> getExperiencesByCandidat(Long candidatId) {
-        return experienceRepository.findByCandidatId(candidatId).stream()
+        return savedExperiences.stream()
                 .map(ExperienceMapper::toDTO)
                 .toList();
     }
 
+    @Override
+    public List<ExperienceDTO> getExperiencesByCandidat(Long candidatId) throws CandidatNotFoundException {
+        if (!candidatRepository.existsById(candidatId)) {
+            throw new CandidatNotFoundException("Candidat non trouvé avec id " + candidatId);
+        }
+
+        return experienceRepository.findByCandidatId(candidatId)
+                .stream()
+                .map(ExperienceMapper::toDTO)
+                .toList();
+    }
 
     @Override
-    public void deleteExperience(Long id) {
+    public void deleteExperience(Long id) throws ExperienceNotFoundException {
+        if (!experienceRepository.existsById(id)) {
+            throw new ExperienceNotFoundException("Expérience non trouvée avec id " + id);
+        }
         experienceRepository.deleteById(id);
     }
 }
-
